requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.2.0
dnspython==2.3.0
filelock==3.12.2
frozenlist==1.3.3
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
multidict==6.0.4
numpy==1.25.1
openai==0.27.8
packaging==23.1
pandas==2.0.3
pymongo==4.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
safetensors==0.3.1
six==1.16.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
yarl==1.9.2
